Rename .h files to .hpp + formatting fixes
- fix include guards to cpplint style
- fix order of includes to match ros guide
