+++ image_transport/include/image_transport/exception.hpp
@@ -32,56 +32,42 @@
 *  POSSIBILITY OF SUCH DAMAGE.
 *********************************************************************/
 
-#ifndef IMAGE_TRANSPORT_TRANSPORT_HINTS_H
-#define IMAGE_TRANSPORT_TRANSPORT_HINTS_H
+#ifndef IMAGE_TRANSPORT__EXCEPTION_HPP_
+#define IMAGE_TRANSPORT__EXCEPTION_HPP_
 
-#include <memory>
-#include <string>
-
-#include <rclcpp/node.hpp>
+#include <stdexcept>
 
 #include "image_transport/visibility_control.hpp"
 
-namespace image_transport
-{
+namespace image_transport {
 
 /**
- * \brief Stores transport settings for an image topic subscription.
+ * \brief A base class for all image_transport exceptions inheriting from std::runtime_error.
  */
-class TransportHints
+class Exception : public std::runtime_error
 {
 public:
-  /**
-   * \brief Constructor.
-   *
-   * The default transport can be overridden by setting a certain parameter to the
-   * name of the desired transport. By default this parameter is named "image_transport"
-   * in the node's local namespace. For consistency across ROS applications, the
-   * name of this parameter should not be changed without good reason.
-   *
-   * @param node Node to use when looking up the transport parameter.
-   * @param default_transport Preferred transport to use
-   * @param parameter_name The name of the transport parameter
-   */
-  IMAGE_TRANSPORT_PUBLIC
-  TransportHints(
-    const rclcpp::Node * node,
-    const std::string & default_transport = "raw",
-    const std::string & parameter_name = "image_transport")
-  {
-    node->get_parameter_or<std::string>(parameter_name, transport_, default_transport);
-  }
-
-  IMAGE_TRANSPORT_PUBLIC
-  const std::string & getTransport() const
-  {
-    return transport_;
-  }
-
-private:
-  std::string transport_;
+  Exception(const std::string& message) : std::runtime_error(message) {}
 };
 
-} //namespace image_transport
+/**
+ * \brief An exception class thrown when image_transport is unable to load a requested transport.
+ */
+class TransportLoadException : public Exception
+{
+public:
+  TransportLoadException(const std::string& transport, const std::string& message)
+    : Exception("Unable to load plugin for transport '" + transport + "', error string:\n" + message),
+      transport_(transport.c_str())
+  {
+  }
 
-#endif
+  std::string getTransport() const { return transport_; }
+
+protected:
+  const char* transport_;
+};
+
+}  // namespace image_transport
+
+#endif // IMAGE_TRANSPORT__EXCEPTION_HPP_
