+++ image_transport/include/image_transport/raw_subscriber.hpp
@@ -32,67 +32,44 @@
 *  POSSIBILITY OF SUCH DAMAGE.
 *********************************************************************/
 
-#ifndef IMAGE_TRANSPORT_SINGLE_SUBSCRIBER_PUBLISHER
-#define IMAGE_TRANSPORT_SINGLE_SUBSCRIBER_PUBLISHER
-
-#include "rclcpp/macros.hpp"
-#include "sensor_msgs/msg/image.hpp"
+#ifndef IMAGE_TRANSPORT__RAW_SUBSCRIBER_HPP_
+#define IMAGE_TRANSPORT__RAW_SUBSCRIBER_HPP_
 
 #include <sensor_msgs/msg/image.hpp>
 
-#include <string>
-#include <functional>
-
+#include "image_transport/simple_subscriber_plugin.hpp"
 #include "image_transport/visibility_control.hpp"
 
 namespace image_transport {
 
 /**
- * \brief Allows publication of an image to a single subscriber. Only available inside
- * subscriber connection callbacks.
+ * \brief The default SubscriberPlugin.
+ *
+ * RawSubscriber is a simple wrapper for ros::Subscriber which listens for Image messages
+ * and passes them through to the callback.
  */
-class SingleSubscriberPublisher
+class RawSubscriber : public SimpleSubscriberPlugin<sensor_msgs::msg::Image>
 {
-private:
-  SingleSubscriberPublisher(const SingleSubscriberPublisher&) = delete;
-  SingleSubscriberPublisher& operator=( const SingleSubscriberPublisher& ) = delete;
-
 public:
-  typedef std::function<uint32_t()> GetNumSubscribersFn;
-  typedef std::function<void(const sensor_msgs::msg::Image&)> PublishFn;
+  virtual ~RawSubscriber() {}
 
-  IMAGE_TRANSPORT_PUBLIC
-  SingleSubscriberPublisher(
-    const std::string & caller_id, const std::string & topic,
-    const GetNumSubscribersFn & num_subscribers_fn,
-    const PublishFn & publish_fn);
+  virtual std::string getTransportName() const
+  {
+    return "raw";
+  }
 
-  IMAGE_TRANSPORT_PUBLIC
-  std::string getSubscriberName() const;
+protected:
+  virtual void internalCallback(const std::shared_ptr<const sensor_msgs::msg::Image>& message, const Callback& user_cb)
+  {
+    user_cb(message);
+  }
 
-  IMAGE_TRANSPORT_PUBLIC
-  std::string getTopic() const;
-
-  IMAGE_TRANSPORT_PUBLIC
-  uint32_t getNumSubscribers() const;
-
-  IMAGE_TRANSPORT_PUBLIC
-  void publish(const sensor_msgs::msg::Image& message) const;
-
-  IMAGE_TRANSPORT_PUBLIC
-  void publish(const sensor_msgs::msg::Image::ConstSharedPtr& message) const;
-
-private:
-  std::string caller_id_;
-  std::string topic_;
-  GetNumSubscribersFn num_subscribers_fn_;
-  PublishFn publish_fn_;
-
-  friend class Publisher; // to get publish_fn_ directly
+  virtual std::string getTopicToSubscribe(const std::string& base_topic) const
+  {
+    return base_topic;
+  }
 };
 
-typedef std::function<void (const SingleSubscriberPublisher &)> SubscriberStatusCallback;
+}  // namespace image_transport
 
-} //namespace image_transport
-
-#endif
+#endif  // IMAGE_TRANSPORT__RAW_SUBSCRIBER_HPP_
