+++ image_transport/include/image_transport/simple_publisher_plugin.hpp
@@ -32,14 +32,14 @@
 *  POSSIBILITY OF SUCH DAMAGE.
 *********************************************************************/
 
-#ifndef IMAGE_TRANSPORT_SIMPLE_PUBLISHER_PLUGIN_H
-#define IMAGE_TRANSPORT_SIMPLE_PUBLISHER_PLUGIN_H
+#ifndef IMAGE_TRANSPORT__SIMPLE_PUBLISHER_PLUGIN_HPP_
+#define IMAGE_TRANSPORT__SIMPLE_PUBLISHER_PLUGIN_HPP_
 
 #include <rclcpp/node.hpp>
 #include <rclcpp/logger.hpp>
 #include <rclcpp/logging.hpp>
 
-#include "image_transport/publisher_plugin.h"
+#include "image_transport/publisher_plugin.hpp"
 #include "image_transport/visibility_control.hpp"
 
 #include <memory>
@@ -172,6 +172,6 @@ private:
   }
 };
 
-} //namespace image_transport
+}  // namespace image_transport
 
-#endif
+#endif  // IMAGE_TRANSPORT__SIMPLE_PUBLISHER_PLUGIN_HPP_
