+++ image_transport/include/image_transport/simple_subscriber_plugin.hpp
@@ -32,14 +32,14 @@
 *  POSSIBILITY OF SUCH DAMAGE.
 *********************************************************************/
 
-#ifndef IMAGE_TRANSPORT_SIMPLE_SUBSCRIBER_PLUGIN_H
-#define IMAGE_TRANSPORT_SIMPLE_SUBSCRIBER_PLUGIN_H
+#ifndef IMAGE_TRANSPORT__SIMPLE_SUBSCRIBER_PLUGIN_HPP_
+#define IMAGE_TRANSPORT__SIMPLE_SUBSCRIBER_PLUGIN_HPP_
 
 #include <functional>
 #include <memory>
 
 #include "rclcpp/subscription.hpp"
-#include "image_transport/subscriber_plugin.h"
+#include "image_transport/subscriber_plugin.hpp"
 #include "image_transport/visibility_control.hpp"
 
 namespace image_transport
@@ -140,6 +140,6 @@ private:
 
 };
 
-} //namespace image_transport
+}  // namespace image_transport
 
-#endif
+#endif  // IMAGE_TRANSPORT__SIMPLE_SUBSCRIBER_PLUGIN_HPP_
