+++ image_transport/src/list_transports.cpp
@@ -32,11 +32,13 @@
 *  POSSIBILITY OF SUCH DAMAGE.
 *********************************************************************/
 
-#include "image_transport/camera_common.h"
-#include "image_transport/publisher_plugin.h"
-#include "image_transport/subscriber_plugin.h"
-#include <pluginlib/class_loader.hpp>
 #include <map>
+
+#include <pluginlib/class_loader.hpp>
+
+#include "image_transport/camera_common.hpp"
+#include "image_transport/publisher_plugin.hpp"
+#include "image_transport/subscriber_plugin.hpp"
 
 using namespace image_transport;
 using namespace pluginlib;
