+++ image_transport/test/test_message_passing.cpp
@@ -39,8 +39,8 @@
 
 #include <gtest/gtest.h>
 
+#include "image_transport/image_transport.hpp"
 #include "sensor_msgs/msg/image.hpp"
-#include "image_transport/image_transport.h"
 
 #include "utils.hpp"
