+++ image_transport/test/test_publisher.cpp
@@ -5,7 +5,7 @@
 
 #include "rclcpp/rclcpp.hpp"
 
-#include "image_transport/image_transport.h"
+#include "image_transport/image_transport.hpp"
 
 class TestPublisher : public ::testing::Test
 {
