+++ image_transport/test/test_remapping.cpp
@@ -8,7 +8,7 @@
 #include "rclcpp/node.hpp"
 #include "utils.hpp"
 
-#include "image_transport/image_transport.h"
+#include "image_transport/image_transport.hpp"
 
 class TestPublisher : public ::testing::Test
 {
