+++ image_transport/tutorial/include/image_transport_tutorial/resized_subscriber.hpp
@@ -1,4 +1,4 @@
-#include <image_transport/simple_subscriber_plugin.h>
+#include <image_transport/simple_subscriber_plugin.hpp>
 #include <image_transport_tutorial/ResizedImage.h>
 
 class ResizedSubscriber : public image_transport::SimpleSubscriberPlugin<image_transport_tutorial::ResizedImage>
