+++ image_transport/tutorial/src/my_subscriber.cpp
@@ -1,7 +1,7 @@
-#include <ros/ros.h>
-#include <image_transport/image_transport.h>
-#include <opencv2/highgui/highgui.hpp>
 #include <cv_bridge/cv_bridge.h>
+#include <image_transport/image_transport.hpp>
+#include <opencv2/highgui/highgui.hpp>
+#include <ros/ros.h>
 
 void imageCallback(const sensor_msgs::ImageConstPtr& msg)
 {
